pyproject.toml
[tool.poetry]
name = "lupaesg"
version = "0.1.0"
description = ""
authors = ["Paulo Fehlauer <paulofeh@gmail.com>"]
readme = "README.md"
packages = [{include = "lupaesg"}]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "0.109.2"
uvicorn = "0.27.1"
python-dotenv = "1.0.1"
pymongo = "4.6.1"
motor = "3.3.2"
pydantic = "2.5.3"
pydantic-settings = "2.1.0"
boto3 = "^1.34.0"
openai = "^1.52.2"
numpy = "^2.1.2"
pinecone-client = "^5.0.1"

[tool.poetry.group.dev.dependencies]
black = "24.1.1"
pylint = "3.0.3"
flake8 = "7.0.0"
pytest = "^8.0.0"
pytest-cov = "^4.1.0"
pytest-asyncio = "^0.23.5"
ipykernel = "^6.29.0"

[tool.pylint.messages_control]
disable = [
    "C0111",  # missing-docstring
    "C0103",  # invalid-name
    "C0330",  # bad-continuation
]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --cov=lupaesg"
testpaths = ["tests"]

[tool.black]
line-length = 100
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"